for null login token, change root user path to public
The original behavior left the default, which is to write to c\users\default -- this is not a writable location, and so the hub is useless. (I found this by testing against the TmpAuthenticator). It may not be ideal, but changing it to public documents at least makes it usable. 

I'm less familiar with python so I opted for more checks to avoid exceptions. There is probably a clearer way to write the change.